+++ R/ddml_late.R
@@ -55,10 +55,16 @@
 #'     Omission of the \code{args} element results in default arguments being
 #'     used in \code{fun}. Omission of \code{assign_X} (and/or \code{assign_Z})
 #'     results in inclusion of all variables in \code{X} (and/or \code{Z}).
-#' @param learners_DXZ Optional argument to allow for different estimators of
-#'     \eqn{E[D \vert X, Z]}. Setup is identical to \code{learners}.
-#' @param learners_ZX Optional argument to allow for different estimators of
-#'     \eqn{E[Z\vert X]}. Setup is identical to \code{learners}.
+#' @param learners_DXZ,learners_ZX Optional arguments to allow for different
+#'     estimators of \eqn{E[D \vert X, Z]}, \eqn{E[Z \vert X]}. Setup is
+#'     identical to \code{learners}.
+#' @param custom_ensemble_weights_DXZ,custom_ensemble_weights_ZX Optional
+#'     arguments to allow for different
+#'     custom ensemble weights for \code{learners_DXZ},\code{learners_ZX}. Setup
+#'     is identical to \code{custom_ensemble_weights}. Note:
+#'     \code{custom_ensemble_weights} and
+#'     \code{custom_ensemble_weights_DXZ},\code{custom_ensemble_weights_ZX} must
+#'     have the same number of columns.
 #' @param subsamples_Z0,subsamples_Z1 List of vectors with sample indices for
 #'     cross-fitting, corresponding to observations with \eqn{Z=0} and
 #'     \eqn{Z=1}, respectively.
@@ -119,36 +125,42 @@
 #' summary(late_fit)
 #'
 #' # Estimate the local average treatment effect using short-stacking with base
-#' #     learners ols, lasso, and ridge.
+#' #     learners ols, lasso, and ridge. We can also use custom_ensemble_weights
+#' #     to estimate the ATE using every individual base learner.
+#' weights_everylearner <- diag(1, 3)
+#' colnames(weights_everylearner) <- c("mdl:ols", "mdl:lasso", "mdl:ridge")
 #' late_fit <- ddml_late(y, D, Z, X,
 #'                       learners = list(list(fun = ols),
 #'                                       list(fun = mdl_glmnet),
 #'                                       list(fun = mdl_glmnet,
 #'                                            args = list(alpha = 0))),
 #'                       ensemble_type = 'nnls',
+#'                       custom_ensemble_weights = weights_everylearner,
 #'                       shortstack = TRUE,
 #'                       sample_folds = 2,
 #'                       silent = TRUE)
 #' summary(late_fit)
 ddml_late <- function(y, D, Z, X,
-                     learners,
-                     learners_DXZ = learners,
-                     learners_ZX = learners,
-                     sample_folds = 2,
-                     ensemble_type = "nnls",
-                     shortstack = FALSE,
-                     cv_folds = 5,
-                     subsamples_Z0 = NULL,
-                     subsamples_Z1 = NULL,
-                     cv_subsamples_list_Z0 = NULL,
-                     cv_subsamples_list_Z1 = NULL,
-                     silent = FALSE) {
+                      learners,
+                      learners_DXZ = learners,
+                      learners_ZX = learners,
+                      sample_folds = 2,
+                      ensemble_type = "nnls",
+                      shortstack = FALSE,
+                      cv_folds = 5,
+                      custom_ensemble_weights = NULL,
+                      custom_ensemble_weights_DXZ = custom_ensemble_weights,
+                      custom_ensemble_weights_ZX = custom_ensemble_weights,
+                      subsamples_Z0 = NULL,
+                      subsamples_Z1 = NULL,
+                      cv_subsamples_list_Z0 = NULL,
+                      cv_subsamples_list_Z1 = NULL,
+                      silent = FALSE) {
   # Data parameters
   nobs <- length(y)
   is_Z0 <- which(Z == 0)
   nobs_Z0 <- length(is_Z0)
   nobs_Z1 <- nobs - nobs_Z0
-  nensb <- length(ensemble_type)
 
   # Create sample fold tuple by treatment
   if (is.null(subsamples_Z0) | is.null(subsamples_Z1)) {
@@ -200,6 +212,7 @@ ddml_late <- function(y, D, Z, X,
   y_X_Z0_res <- get_CEF(y[is_Z0], X[is_Z0, , drop = F],
                         learners = learners, ensemble_type = ensemble_type,
                         shortstack = shortstack,
+                        custom_ensemble_weights = custom_ensemble_weights,
                         cv_subsamples_list = cv_subsamples_list_Z0,
                         subsamples = subsamples_Z0,
                         silent = silent, progress = "E[Y|Z=0,X]: ",
@@ -209,39 +222,67 @@ ddml_late <- function(y, D, Z, X,
   y_X_Z1_res <- get_CEF(y[-is_Z0], X[-is_Z0, , drop = F],
                         learners = learners, ensemble_type = ensemble_type,
                         shortstack = shortstack,
+                        custom_ensemble_weights = custom_ensemble_weights,
                         cv_subsamples_list = cv_subsamples_list_Z1,
                         subsamples = subsamples_Z1,
                         silent = silent, progress = "E[Y|Z=1,X]: ",
                         auxilliary_X = auxilliary_X_Z0)
 
-  # Compute estimates of E[D|Z=0,X]
-  D_X_Z0_res <- get_CEF(D[is_Z0], X[is_Z0, , drop = F],
-                        learners = learners_DXZ,
-                        ensemble_type = ensemble_type,
-                        shortstack = shortstack,
-                        cv_subsamples_list = cv_subsamples_list_Z0,
-                        subsamples = subsamples_Z0,
-                        silent = silent, progress = "E[D|Z=0,X]: ",
-                        auxilliary_X = auxilliary_X_Z1)
+  # Check for perfect non-compliance
+  if (all(D[Z==0] == 0)) {
+    # Artificially construct values for subsample with Z=0
+    D_X_Z0_res <- list(NULL)
+    D_X_Z0_res$oos_fitted <- rep(0, nobs_Z0)
+    D_X_Z0_res$auxilliary_fitted <-
+      lapply(y_X_Z0_res$auxilliary_fitted, function (x) {x * 0})
+    if (!silent) cat("E[D|Z=0,X]: perfect non-compliance -- Done! \n")
+  } else {
+    # Compute estimates of E[D|Z=0,X]
+    D_X_Z0_res <- get_CEF(D[is_Z0], X[is_Z0, , drop = F],
+                          learners = learners_DXZ,
+                          ensemble_type = ensemble_type,
+                          shortstack = shortstack,
+                          custom_ensemble_weights = custom_ensemble_weights_DXZ,
+                          cv_subsamples_list = cv_subsamples_list_Z0,
+                          subsamples = subsamples_Z0,
+                          silent = silent, progress = "E[D|Z=0,X]: ",
+                          auxilliary_X = auxilliary_X_Z1)
+  }#IFELSE
 
-  # Compute estimates of E[D|Z=1,X]
-  D_X_Z1_res <- get_CEF(D[-is_Z0], X[-is_Z0, , drop = F],
-                        learners = learners_DXZ,
-                        ensemble_type = ensemble_type,
-                        shortstack = shortstack,
-                        cv_subsamples_list = cv_subsamples_list_Z1,
-                        subsamples = subsamples_Z1,
-                        silent = silent, progress = "E[D|Z=1,X]: ",
-                        auxilliary_X = auxilliary_X_Z0)
+  # Check for perfect compliance
+  if (all(D[Z==1] == 1)) {
+    # Artificially construct values for subsample with Z=0
+    D_X_Z1_res <- list(NULL)
+    D_X_Z1_res$oos_fitted <- rep(0, nobs_Z1)
+    D_X_Z1_res$auxilliary_fitted <-
+      lapply(y_X_Z1_res$auxilliary_fitted, function (x) {x * 0})
+    if (!silent) cat("E[D|Z=1,X]: perfect compliance -- Done! \n")
+  } else {
+    # Compute estimates of E[D|Z=1,X]
+    D_X_Z1_res <- get_CEF(D[-is_Z0], X[-is_Z0, , drop = F],
+                          learners = learners_DXZ,
+                          ensemble_type = ensemble_type,
+                          shortstack = shortstack,
+                          custom_ensemble_weights = custom_ensemble_weights_DXZ,
+                          cv_subsamples_list = cv_subsamples_list_Z1,
+                          subsamples = subsamples_Z1,
+                          silent = silent, progress = "E[D|Z=1,X]: ",
+                          auxilliary_X = auxilliary_X_Z0)
+  }#IFELSE
 
   # Compute estimates of E[Z|X]
   Z_X_res <- get_CEF(Z, X,
                      learners = learners_ZX, ensemble_type = ensemble_type,
                      shortstack = shortstack,
+                     custom_ensemble_weights = custom_ensemble_weights_ZX,
                      cv_subsamples_list = cv_subsamples_list,
                      subsamples = subsamples,
                      compute_insample_predictions = F,
                      silent = silent, progress = "E[Z|X]: ")
+
+  # Update ensemble type to account for (optional) custom weights
+  ensemble_type <- dimnames(y_X_Z0_res$weights)[[2]]
+  nensb <- ifelse(is.null(ensemble_type), 1, length(ensemble_type))
 
   # Check whether multiple ensembles are computed simultaneously
   multiple_ensembles <- nensb > 1