+++ tests/testthat/test-ddml_att.R
@@ -1,4 +1,4 @@
-test_that("ddml_ate computes with a single model", {
+test_that("ddml_att computes with a single model", {
   # Simulate small dataset
   nobs <- 200
   X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
@@ -7,16 +7,18 @@ test_that("ddml_ate computes with a single model", {
   y <- D + X %*% runif(40) + rnorm(nobs)
   # Define arguments
   learners <- list(what = ols)
-  ddml_ate_fit <- ddml_ate(y, D, X,
+  learners_DX <- list(what = mdl_glm)
+  ddml_att_fit <- ddml_att(y, D, X,
                            learners = learners,
+                           learners_DX = learners_DX,
                            cv_folds = 3,
                            sample_folds = 3,
                            silent = T)
   # Check output with expectations
-  expect_equal(length(ddml_ate_fit$ate), 1)
+  expect_equal(length(ddml_att_fit$att), 1)
 })#TEST_THAT
 
-test_that("ddml_ate computes with an ensemble procedure", {
+test_that("ddml_att computes with an ensemble procedure", {
   # Simulate small dataset
   nobs <- 200
   X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
@@ -27,17 +29,17 @@ test_that("ddml_ate computes with an ensemble procedure", {
   learners <- list(list(fun = ols),
                    list(fun = ols))
   # Compute DDML PLM estimator
-  ddml_ate_fit <- ddml_ate(y, D, X,
+  ddml_att_fit <- ddml_att(y, D, X,
                            learners = learners,
                            ensemble_type = "ols",
                            cv_folds = 3,
                            sample_folds = 3,
                            silent = T)
   # Check output with expectations
-  expect_equal(length(ddml_ate_fit$ate), 1)
+  expect_equal(length(ddml_att_fit$att), 1)
 })#TEST_THAT
 
-test_that("ddml_ate computes w/ multiple ensembles + custom weights", {
+test_that("ddml_att computes w/ multiple ensembles + custom weights", {
   # Simulate small dataset
   nobs <- 200
   X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
@@ -48,7 +50,7 @@ test_that("ddml_ate computes w/ multiple ensembles + custom weights", {
   learners <- list(list(fun = ols),
                    list(fun = ols))
   # Compute DDML PLM estimator
-  ddml_ate_fit <- ddml_ate(y, D, X,
+  ddml_att_fit <- ddml_att(y, D, X,
                            learners,
                            ensemble_type = c("ols", "nnls",
                                              "singlebest", "average"),
@@ -57,10 +59,10 @@ test_that("ddml_ate computes w/ multiple ensembles + custom weights", {
                            sample_folds = 3,
                            silent = T)
   # Check output with expectations
-  expect_equal(length(ddml_ate_fit$ate), 6)
+  expect_equal(length(ddml_att_fit$att), 6)
 })#TEST_THAT
 
-test_that("ddml_ate computes with multiple ensemble procedures & shortstack", {
+test_that("ddml_att computes w/ multp ensembles, custom weights + shortstack", {
   # Simulate small dataset
   nobs <- 200
   X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
@@ -68,21 +70,22 @@ test_that("ddml_ate computes with multiple ensemble procedures & shortstack", {
   D <- 1 * (D_tld > mean(D_tld))
   y <- D + X %*% runif(40) + rnorm(nobs)
   # Define arguments
-  learners <- list(list(fun = ols))
+  learners <- list(list(fun = ols),
+                   list(fun = ols))
   # Compute DDML PLM estimator
-  ddml_ate_fit <- ddml_ate(y, D, X,
+  ddml_att_fit <- ddml_att(y, D, X,
                            learners,
-                           ensemble_type = c("ols", "nnls",
-                                             "singlebest", "average"),
+                           ensemble_type = c("ols", "average"),
                            shortstack = TRUE,
                            cv_folds = 3,
+                           custom_ensemble_weights = diag(1, 2),
                            sample_folds = 3,
                            silent = T)
   # Check output with expectations
-  expect_equal(length(ddml_ate_fit$ate), 4)
+  expect_equal(length(ddml_att_fit$att), 4)
 })#TEST_THAT
 
-test_that("summary.ddml_ate computes with a single model", {
+test_that("summary.ddml_att computes with a single model", {
   # Simulate small dataset
   nobs <- 200
   X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
@@ -91,17 +94,17 @@ test_that("summary.ddml_ate computes with a single model", {
   y <- D + X %*% runif(40) + rnorm(nobs)
   # Define arguments
   learners <- list(what = ols)
-  ddml_ate_fit <- ddml_ate(y, D, X,
+  ddml_att_fit <- ddml_att(y, D, X,
                            learners = learners,
                            cv_folds = 3,
                            sample_folds = 3,
                            silent = T)
-  capture_output({inf_res <- summary(ddml_ate_fit)})
+  capture_output({inf_res <- summary(ddml_att_fit)})
   # Check output with expectations
   expect_equal(length(inf_res), 4)
 })#TEST_THAT
 
-test_that("summary.ddml_ate computes with multiple ensemble procedures", {
+test_that("summary.ddml_att computes with multiple ensemble procedures", {
   # Simulate small dataset
   nobs <- 200
   X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
@@ -111,14 +114,14 @@ test_that("summary.ddml_ate computes with multiple ensemble procedures", {
   # Define arguments
   learners <- list(list(fun = ols))
   # Compute DDML PLM estimator
-  ddml_ate_fit <- ddml_ate(y, D, X,
+  ddml_att_fit <- ddml_att(y, D, X,
                            learners,
                            ensemble_type = c("ols", "nnls",
                                              "singlebest", "average"),
                            cv_folds = 3,
                            sample_folds = 3,
                            silent = T)
-  capture_output({inf_res <- summary(ddml_ate_fit)}, print = FALSE)
+  capture_output({inf_res <- summary(ddml_att_fit)}, print = FALSE)
   # Check output with expectations
   expect_equal(length(inf_res), 16)
 })#TEST_THAT