+++ NEWS.md
@@ -1,3 +1,11 @@
+# ddml 0.2.0
+
+* Adds support for the average treatment effect on the treated estimator.
+* Adds support for local average treatment effect estimation with perfect compliance or perfect non-compliance.
+* Adds support for custom ensemble weights.
+* Adds article on integration with the ``did`` package.
+* Adds ``mdl_glm`` wrapper for ``stats::glm()``.
+
 # ddml 0.1.0
 
 * Initial CRAN submission.